requirements.txt
django
djangorestframework
markdown
django-filter
requests
django-jsonfield